Scrap Code 1
DONT TRY TO COMPILE AS I HAVE NOT ADDED REQUIRED BINARIES. JUST REVIEW and provide key things to consider.

Post Master base is the base class from where the PostMasterSendGrid and PostMasterMailGun are derived and which has the SendMail logic for each mailing service. Setting.xml needs to be integrated in the Init such that configurable constants for MailGun and SendGrid is read from it.

Mailer.cs implements IMailer interface and has the main logic for data cleansing, verification, failsafe node selection etc.